Document modular architecture
